chargen: Adjustments, support for changing stats, alignment, hp
* Only SDL_BUTTON_LEFT or SDL_BUTTON_RIGHT mouseclicks are recognized now
* mouse_up events will only register if a mouse_down event was fired within the bounds of the target sprite (holding left-click over the race portrait and releasing it over the die will no longer roll the die)
* Added buttons for stats/alignment/hp. Still need to add constraints - no logic behind this aside from ++ and --
* Added name text box, can capture and lose focus as well. The text cursor used is different from the original, but is an actual text cursor instead of a yellow 'l' or something. Not sure what was used for that, but we can change it match the original.

FIXME: I don't think I'm using the proper graphic for the buttons for the stats/alignment/hp fields. I thought it the image id was 20047 at frame 0, but the color is wrong and swapping palettes didn't get me to the color I was looking for. Did they manually change the color in the code or am I doing something wrong? It looks like they have a shadow (probably just a black version of the same image, which you can palette out) of each button behind the buttons in character creation. Fix the text cursor as well, if we want to match the original.